LayoutTests/fast/forms/ValidityState-stepMismatch-number-expected.txt

Check stepMismatch results for type=number.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is true
PASS input.validity.stepMismatch is true
PASS input.validity.stepMismatch is true
PASS input.value is "5"
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is true
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is true
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS successfullyParsed is true

TEST COMPLETE
